third_party/WebKit/LayoutTests/battery-status/multiple-windows-expected.txt

Test battery status API with multiple windows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


resolution in window
PASS batteryInfo is defined.
PASS mockBatteryInfo is defined.
PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
resolution in secondWindow
PASS batteryInfo is defined.
PASS mockBatteryInfo is defined.
PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
PASS successfullyParsed is true

TEST COMPLETE
